moved elevation import code to StreetMapImporting Module
