+++ Source/StreetMapImporting/Elevation.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "Landscape.h"
+
+ALandscape* SpawnLandscape(UWorld* World);
